chore: refresh arraybuffer inspect CI metadata
